pyproject.toml
[tool.poetry]
name = "familiarity-playlist-backend"
version = "0.1.0"
description = ""
authors = ["Pony0w0 <pony8674312@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.10.13"
django = "5.2.2"
django-redis = "5.0.0"
djangorestframework = "3.15.2"
django-filter = "24.3"
django-cryptography = "1.1"
django-extensions = "4.1"
psycopg2-binary = "2.9.6"
python-dotenv = "1.0.0"
django-cors-headers = "4.6.0"
ipython = "8.10"
gunicorn = "23.0.0"
djangorestframework-simplejwt = "5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
